requirements.txt
crewai==0.22.2
python-dotenv
openai>=1.0.0
streamlit
# For PDF handling
PyPDF2
# For voice input
SpeechRecognition
pyaudio
